// capacitor/plugins-library/src/opener/models/KrikooMimeType.ts
export enum KrikooMimeType {
  BmpImage = 'image/bmp',
  Document = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  GifImage = 'image/gif',
  JpegImage = 'image/jpeg',
  MSDocument = 'application/msword',
  MSPresentation = 'application/vnd.ms-powerpoint',
  MSSheet = 'application/vnd.ms-excel',
  Pdf = 'application/pdf',
  PlainText = 'text/plain',
  PngImage = 'image/png',
  Presentation = 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
  Sheet = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  TifImage = 'image/tif'
}